+++ src/templates/certificateOfAchievement/CertificateOfAchievement.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Preview, Props, Description, Story } from "@storybook/addon-docs/blocks";
 import { object } from "@storybook/addon-knobs";
 import { CertificateOfAchievement } from "./certificateOfAchievement";
-import { sampleCertificate } from "./sample";
+import { sampleCertificate } from "./fixtures/sample";
 
 <Meta title="MDX|CertificateOfAchievementTemplate" component={CertificateOfAchievement} />
