+++ src/templates/certificateOfAchievement/recommendations.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Preview, Props, Description, Story } from "@storybook/addon-docs/blocks";
 import { object } from "@storybook/addon-knobs";
 import { Recommendations } from "./recommendations";
-import { sampleCertificate } from "./sample";
+import { sampleCertificate } from "./fixtures/sample";
 
 <Meta title="MDX|CertificateOfAchievementTemplate" component={Recommendations} />
