+++ src/templates/certificateOfAchievement/utils/styles.tsx
@@ -1,0 +1,10 @@
+// once opencerts adhere to bootstrap's container class, max-width will need to change to align
+export const containerCustom = (): string => {
+  return `
+    font-family: "Lucida Sans Unicode", "Lucida Grande", sans-serif;
+      position: relative;
+      max-width: 1280px;
+      margin-left: auto;
+      margin-right: auto;
+    `;
+};
