+++ src/templates/index.tsx
@@ -1,6 +1,6 @@
 import { TemplateRegistry } from "@govtechsg/decentralized-renderer-react-components";
 import { templates } from "./certificateOfAchievement";
-import { GovTechCertificateTemplate } from "./certificateOfAchievement/sample";
+import { GovTechCertificateTemplate } from "./certificateOfAchievement/types";
 
 export const registry: TemplateRegistry<GovTechCertificateTemplate> = {
   CERTIFICATE_OF_ACHIEVEMENT: templates
